chore: use getAddress instead of getAddresses function
