+++ src/extension/background-script/router.ts
@@ -60,7 +60,7 @@ const routes = {
   lnurlAuth: auth,
   getCurrencyRate: cache.getCurrencyRate,
   signPsbt: webbtc.signPsbt,
-  getAddresses: webbtc.getAddresses,
+  getAddress: webbtc.getAddress,
   setMnemonic: mnemonic.setMnemonic,
   getMnemonic: mnemonic.getMnemonic,
 
@@ -77,7 +77,7 @@ const routes = {
       enable: allowances.enable,
       getInfo: webbtc.getInfo,
       signPsbtWithPrompt: webbtc.signPsbtWithPrompt,
-      getAddressesWithPrompt: webbtc.getAddressesWithPrompt,
+      getAddressWithPrompt: webbtc.getAddressWithPrompt,
     },
     webln: {
       enable: allowances.enable,
